+++ CHANGELOG.md
@@ -7,11 +7,28 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.1.1] - 2020-12-30
+
+### Added
+
+- Evaluator with AST
+
+### Changed
+
+- Decode fn accepts now `&[u8]` instead of `&str`
+
+### Fixed
+
+- Remove space between encoded words [#2]
+
 ## [0.1.0] - 2020-12-28
 
 ### Added
 
 - Init repo
 
-[unreleased]: https://github.com/soywod/rfc2047-decoder/compare/v0.1.0...HEAD
+[unreleased]: https://github.com/soywod/rfc2047-decoder/compare/v0.1.1...HEAD
+[0.1.1]: https://github.com/soywod/rfc2047-decoder/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/soywod/rfc2047-decoder/releases/tag/v0.1.0
+
+[#2]: https://github.com/soywod/rfc2047-decoder/issues/2